% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_comp.R
\name{ar_comp}
\alias{ar_comp}
\title{Function that computes AR for treated firm.}
\usage{
ar_comp(...)
}
\arguments{
\item{data}{The name of the data.table that contains the data.}
}
\value{
A data.table containing the following columns:
\item{tau}{Relative time to event date.}
\item{ar}{Abnormal return.}
\item{sigma}{"Goodness" of fit measure.}
}
\description{
\code{ar_comp} is used calculate the abnormal return for a treated firm. The abnormal return
is computed as the raw return of the treated firm minus the return of its synthetic match.
}
