% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_panel.R
\name{event_panel}
\alias{event_panel}
\title{Function that pre-processes returns of the control group}
\usage{
event_panel(dt_treat, r_control, estwind, eventwind)
}
\arguments{
\item{estwind}{Argument to set estimation window period in relative time to event, i.e. \code{c(estwind_start, estwind_end)}}

\item{eventwind}{Argument to set event window period in relative time to event, i.e. \code{c(eventwind_start, eventwind_end)}}

\item{ed}{date of event.}

\item{data}{The name of the data.table that contains the data.}

\item{estobs_min}{Argument to define minimum number of trading days during the event window. Can be an
integer or a proportional (i.e. between 0 and 1). Default is \eqn{1}, i.e. no missing trading days are allowed.}
}
\value{
A data.table containing the following columns:
\item{cid}{Unique (event \eqn{\times}) control firm identifier.  Treated firm get the unassigned cid of 0.}
\item{t}{Time variable (consecutive).}
\item{estwind}{Estimation window indicator.}
\item{eventwind}{window indicator.}
\item{r}{Stock return.}
}
\description{
\code{get_event_panel} is used to create final "event panel" data set. Control corporations
are chosen such that a balanced panel in relative event time is created.
}
