% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigmafun}
\alias{sigmafun}
\title{Helper Functions used throughout package}
\usage{
sigmafun(x)
}
\description{
Helper Functions used throughout package
}
\section{Functions}{
\itemize{
\item \code{sigmafun()}: of match fit function (equation 8) in Acemoglu et al. (2016).

}}
