% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Helper Functions used throughout package}
\usage{
... \%notin\% NA
}
\description{
Helper Functions used throughout package
}
\section{Functions}{
\itemize{
\item \code{... \%notin\% NA}: of \code{\%in\%}.

}}
