% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_comp.R
\name{phi_comp}
\alias{phi_comp}
\title{Function that computes treatment effect for treatment group.}
\usage{
phi_comp(r_treat, r_control, estwind)
}
\arguments{
\item{r_control}{Returns of control group (includes placebo treatment group returns.}

\item{estwind}{Argument to set estimation window period in relative time to event,
i.e. \code{c(estwind_start, estwind_end)}}

\item{r_treat_placebo}{Returns of placebo treatment group.}
}
\value{
A data.table containing the following columns:
\item{tau}{Relative time to event date.}
\item{phi}{Abnormal return.}
}
\description{
\code{phi_comp} Wrapper function to compute \eqn{\phi} for
placebo treatment group.
}
