% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_comp_placebo.R
\name{phi_comp_placebo}
\alias{phi_comp_placebo}
\title{Function that computes treatment effect for placebo treatment group.}
\usage{
phi_comp_placebo(
  placebo_treat_ids,
  r_control_ed,
  estwind,
  eventwind,
  sigma_cutoff
)
}
\arguments{
\item{estwind}{Argument to set estimation window period in relative time to event,
i.e. \code{c(estwind_start, estwind_end)}}

\item{r_treat_placebo}{Returns of placebo treatment group.}

\item{dt_control}{Returns of control group (includes placebo treatment group returns.}
}
\value{
A data.table containing the following columns:
\item{tau}{Relative time to event date.}
\item{phi}{(Placebo) treatment effect.}
}
\description{
\code{phi_comp_placebo} Wrapper function to compute \eqn{\phi} for
placebo treatment group.
}
