% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_synthReturn.R
\name{pre_process_synthReturn}
\alias{pre_process_synthReturn}
\title{Function that pre-process the data to use with synthReturn}
\usage{
pre_process_synthReturn(
  DT,
  unitname,
  treatname,
  rname,
  dname,
  edname,
  estwind,
  eventwind,
  estobs_min,
  eventobs_min,
  placebo,
  ngroup,
  ndraws,
  ncores,
  is_windows
)
}
\description{
Function that pre-process the data to use with \code{synthReturn}
}
