% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ret.R
\docType{data}
\name{ret_one_evdate}
\alias{ret_one_evdate}
\title{Simulated repeated cross-section data}
\format{
The main data frame \code{ret_two_evdates_na} has 14000 rows and 7 variables:
\describe{
\item{treatid}{unique "firm x event" identifier for each firm in the treatment group.
NA if firm not in treatment group.}
\item{date}{date variable}
\item{ret}{raw stock return}
\item{estwind}{an indicator variable for the estimation window. Equal to 1 if
the trading day is in the estimation window; equal to 0 if the trading day is outside
the estimation window.}
\item{eventwind}{an indicator variable for the event window. Equal to 1 if
the trading day is in the event window; equal to 0 if the trading day is outside
the event window.}
\item{eventdate}{date when the event happened}
\item{treatid}{unique identifier for each firm in the control group.
NA if firm in treatment group.}
}
}
\usage{
ret_one_evdate
}
\description{
\code{sim_rc} contains 3 simulated stock return data sets
}
\keyword{datasets}
