% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ret_two_evdates.R
\docType{data}
\name{ret_two_evdates}
\alias{ret_two_evdates}
\title{Simulated timeseries of stock returns}
\format{
\code{ret_two_evdates} has 14,000 rows and 5 variables:
\describe{
\item{eventdate}{Calendar date when the event happened (\verb{<Date>}).}
\item{date}{Calendar date (\verb{<Date>}).}
\item{ret}{Raw stock return (\verb{<num>}).}
\item{treat}{Binary treatment indicator (\verb{<lgcl>}).}
\item{unit}{Unique identifier for each treated, respectively control firm (\verb{<int>}).}
}
}
\usage{
ret_two_evdates
}
\description{
\code{ret_two_evdates} presents a simulated timeseries of stock returns for 20
treated firms and a set of 60 potential control firms. The simulated data allows for a max
estimation window of 100 trading days (in event time: -100 to -1) and a max event window
of 6 trading days (in event time: 0 to +5).
}
\keyword{datasets}
