% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_control_set.R
\name{get_control_set}
\alias{get_control_set}
\title{Function that pre-processes returns of the control group}
\usage{
get_control_set(
  eventdate,
  cdata,
  estwind,
  eventwind,
  estobs_min,
  eventobs_min,
  ncontrol_min,
  not_permutation
)
}
\arguments{
\item{eventdate}{date of event.}

\item{estwind}{Argument to set estimation window period in relative time to event, i.e. \code{c(estwind_start, estwind_end)}}

\item{eventwind}{Argument to set event window period in relative time to event, i.e. \code{c(eventwind_start, eventwind_end)}}

\item{estobs_min}{Argument to define minimum number of trading days during the event window. Can be an
integer or a proportional (i.e. between 0 and 1). Default is \eqn{1}, i.e. no missing trading days are allowed.}

\item{data}{The name of the data.table that contains the data.}
}
\value{
A data.table containing the following columns:
\item{cid}{Unique (event \eqn{\times}) control firm identifier.}
\item{d}{Trading Date.}
\item{ed}{Event date.}
\item{r}{Stock return.}
}
\description{
\code{get_treat_set} is used to create an "event panel" of potential control group
firms for each (unique) event date.
}
