% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_comp.R
\name{phi_comp}
\alias{phi_comp}
\title{Function that computes treatment effect for treatment group.}
\usage{
phi_comp(r_treat, r_control)
}
\arguments{
\item{r_control}{Returns of control group (includes placebo treatment group returns.}

\item{r_treat_placebo}{Returns of placebo treatment group.}
}
\value{
A list of 2 data.tables, with the main data.table \code{phi} containing the following columns:
\item{tau}{Relative time to event day.}
\item{phi}{Average treatment effect.}
}
\description{
\code{phi_comp} Wrapper function to compute \eqn{\phi} for
placebo treatment group.
}
