% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthReturn.R
\name{synthReturn}
\alias{synthReturn}
\title{Function that}
\usage{
synthReturn(
  data,
  tidname,
  cidname,
  dname,
  rname,
  edname,
  estwind,
  eventwind,
  estobs_min = 1,
  eventobs_min = 1,
  placebo = TRUE,
  ngroup = 2,
  ndraws = 25,
  ncores = NULL
)
}
\arguments{
\item{data}{The name of the data.frame that contains the data. Data should be stored in the "long" format.}

\item{tidname}{The name of the column containing the treated unit id.}

\item{cidname}{The name of the column containing the control unit id.}

\item{dname}{The name of the column containing the date variable.}

\item{rname}{The name of the column containing the stock returns.}

\item{edname}{The name of the column containing the (treatment firm-specific) event date.}

\item{estwind}{Argument to set estimation window period in relative time to event, i.e. \code{c(estwind_start, estwind_end)}.}

\item{eventwind}{Argument to set event window period in relative time to event, i.e. \code{c(eventwind_start, eventwind_end)}.}

\item{estobs_min}{Argument to define minimum number of trading days during the event window. Can be an
integer or a proportion (i.e. between 0 and 1). Default is \eqn{1}, i.e. no missing trading days are allowed.}

\item{placebo}{Whether inference via placebo treatment group effects should be drawn. Default is \code{TRUE}.}

\item{ngroup}{Minimum number of control firms in placebo event(-date) panel relative to placebo treatment group size.
Default is \eqn{2}, i.e. placebo control group size has to be at least as large as size of placebo treatment group.}

\item{ndraws}{Number of randomly drawn placebo treatment groups at each (unique) event date. Has to be larger than \eqn{1}.}

\item{ncores}{Number of CPU cores to use. \code{NULL} (the default) sets it to the number of available cores.}
}
\value{
An S3 object containing the following components:
\item{ate}{Data.frame containing the average treatment effect estimates \eqn{\phi} and (if \code{placebo == TRUE}) the 90\%, 95\% and 99\% confidence intervals.}
\item{ar}{Data.frame containing the estimated abnormal returns, and the "goodness" of the synthetic match estimate \eqn{\sigma} for all firms in the (actual) treatment group.}
\item{placebo}{List containing the average treatment effect estimates \eqn{\phi} for each placebo treatment group and the number of placebo treatment groups.}
\item{argu}{Some arguments used in the call (estwind, eventwind, estobs_min, eventobs_min, ngroup, ndraws).}
}
\description{
\code{synthReturn} computes the average treatment effect \eqn{\phi} using the synthetic
matching method suggested by Acemoglu et al. (2016) and modified by Kreitmeir et al. (2023) to
accommodate multiple event dates and missing values.
}
\examples{
# -----------------------------------------------
# Example with two event-dates and no missing values
# -----------------------------------------------

# Load data in that comes in the synthReturn package
data(ret_two_evdates)

# -----------------------------------------------
# Implement the synthetic matching matching method

synthReturn(
  data = ret_two_evdates,
  tidname = "treatid",
  cidname = "controlid",
  rname = "ret",
  dname = "date",
  edname = "eventdate",
  estwind = c(-100,-1),
  eventwind = c(0,5),
  estobs_min = 1,
  eventobs_min = 1,
  placebo = TRUE,
  ngroup = 2,
  ndraws = 10
  )


# -----------------------------------------------
# Example with two event-dates and no missing values
# -----------------------------------------------

# Load data in that comes in the synthReturn package
data(ret_two_evdates_na)

# -----------------------------------------------
# Implement the synthetic matching matching method

# Note: You can set a threshold for the minimum of non-missing trading days
# during both the estimation (estobs_min) and event window (eventobs_min). Here, a firm
# is required to have non-missing returns for at least 90\% of trading days during both,
# the estimation and event window (Default is 1 = 100\%).

synthReturn(
  data = ret_two_evdates,
  tidname = "treatid",
  cidname = "controlid",
  rname = "ret",
  dname = "date",
  edname = "eventdate",
  estwind = c(-100,-1),
  eventwind = c(0,5),
  estobs_min = 0.9,
  eventobs_min = 0.9,
  placebo = TRUE,
  ngroup = 2,
  ndraws = 10
)

}
