% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthReturn.R
\name{synthReturn}
\alias{synthReturn}
\title{Function that}
\usage{
synthReturn(
  data,
  tidname,
  cidname,
  dname,
  rname,
  edname,
  estwind,
  eventwind,
  estobs_min = 1,
  eventobs_min = 1,
  placebo = TRUE,
  ndraws = 25,
  ngroup = 2
)
}
\arguments{
\item{data}{The name of the data.frame that contains the data. Data should be stored in the "long" format.}

\item{tidname}{The name of the column containing the control unit id.}

\item{dname}{The name of the column containing the date variable.}

\item{rname}{The name of the column containing the stock returns.}

\item{edname}{The name of the column containing the (treatment firm specific) event date.}

\item{estwind}{Argument to set estimation window period in relative time to event, i.e. \code{c(estwind_start, estwind_end)}}

\item{eventwind}{Argument to set event window period in relative time to event, i.e. \code{c(eventwind_start, eventwind_end)}}

\item{estobs_min}{Argument to define minimum number of trading days during the event window. Can be an
integer or a proportional (i.e. between 0 and 1). Default is \eqn{1}, i.e. no missing trading days are allowed.}

\item{placebo}{Whether inference via placebo treatment group effects should be drawn. Default is \code{TRUE}}

\item{ndraws}{Number of randomly drawn placebo treatment group at each (unique) event date. Has to be larger than \eqn{1}.}

\item{ngroup}{Minimum number of control firms in placebo event(-date) panel relative to placebo treatment group size.
Default is \eqn{2}, i.e. placebo control group size has to be at least as large as size of placebo treatment group.}
}
\value{
A data.table containing the following components:
\item{tau}{Relative time to event day (\eqn{\tau} = 0).}
\item{phi}{The treatment effect estimate for the period \eqn{[0,k]} during the event window.}
\item{ci_90_lower}{90\% confidence interval lower bound; not relevant if placebo = \code{FALSE}.}
\item{ci_90_upper}{90\% confidence interval upper bound; not relevant if placebo = \code{FALSE}.}
\item{ci_95_lower}{95\% confidence interval lower bound; not relevant if placebo = \code{FALSE}.}
\item{ci_95_upper}{95\% confidence interval upper bound; not relevant if placebo = \code{FALSE}.}
\item{ci_99_lower}{99\% confidence interval lower bound; not relevant if placebo = \code{FALSE}.}
\item{ci_99_upper}{99\% confidence interval upper bound; not relevant if placebo = \code{FALSE}.}
\item{n_placebo}{Number of placebo treatment effects; not relevant if placebo = \code{FALSE}.}
}
\description{
\code{synthReturn} computes the treatment effect \eqn{\phi} using the synthetic
matching method suggested by Acemoglu et al. (2016) and modified by Kreitmeir et al. (2023) to
accommodate multiple event dates and missing values.
}
